feat: test cleanup
