+++ pickup/test/_compose.js
@@ -64,7 +64,8 @@ export async function compose () {
     ipfsApiUrl,
     dynamoEndpoint,
     dynamoClient,
-    dynamoTable
+    dynamoTable,
+    sqsContainer
   }
 }
